+++ events/market_updated_event.go
@@ -55,6 +55,7 @@ func (m MarketUpdated) StreamMessage() *eventspb.BusEvent {
 		Version: eventspb.Version,
 		Id:      m.eventID(),
 		Block:   m.TraceID(),
+		ChainId: m.ChainID(),
 		Type:    m.et.ToProto(),
 		Event: &eventspb.BusEvent_MarketUpdated{
 			MarketUpdated: &market,