+++ events/trade.go
@@ -41,6 +41,7 @@ func (t Trade) StreamMessage() *eventspb.BusEvent {
 		Version: eventspb.Version,
 		Id:      t.eventID(),
 		Block:   t.TraceID(),
+		ChainId: t.ChainID(),
 		Type:    t.et.ToProto(),
 		Event: &eventspb.BusEvent_Trade{
 			Trade: &t.t,