+++ events/vote.go
@@ -66,6 +66,7 @@ func (v Vote) StreamMessage() *eventspb.BusEvent {
 		Version: eventspb.Version,
 		Id:      v.eventID(),
 		Block:   v.TraceID(),
+		ChainId: v.ChainID(),
 		Type:    v.et.ToProto(),
 		Event: &eventspb.BusEvent_Vote{
 			Vote: &v.v,